myshop/requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
defusedxml==0.7.1
Django==6.0.2
django-ckeditor-5==0.2.19
django-jazzmin==3.0.1
django-shopping-cart==0.1
idna==3.11
oauthlib==3.3.1
pillow==12.1.0
pycparser==3.0
PyJWT==2.11.0
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.5
urllib3==2.6.3
